Fix ydn-db#90: traverse() call func parameter even if tree is empty.
In traverse, and reverse Traverse methods, on an empty tree, or if there
is no node with value >= to the start value, the func parameter is not called.
In the scan method in the same context, func method is called.
In ydn.db.core.req.SimpleCursor, if the func parameter of buffer is not
called the onSuccess callback is never called and the result state is
never set to 'ready'.
This fix call func function with a null parameter if tree is empty or if
there is no node with value >= to the start value.